--- .gitignore
@@ -1,8 +1,0 @@
-node_modules/
-public/bundle*
-forever/node_modules*
-inject.log
-core
-logs/
-
-apikey
