--- build_names
@@ -1,25 +1,0 @@
-ld-2.26.3.4so
-libbsd.so.0.8.6.1.1
-libc-2.26.3.so
-libcom_err.so.2.1.1
-libcrypt-2.26.1.so
-libdl-2.26.1.so
-libexpat.so.1.6.5.1
-libgcc_s.so.1.1
-libgcrypt.so.20.1.8.1
-libgpg-error.so.0.22.0.3
-libidn.so.11.6.16.5
-libkeyutils.so.1.5.7
-liblzma.so.5.2.2.3
-libm-2.26.4.so
-libnsl-2.26.6.so
-libpcre.so.3.13.3.5
-libpthread-2.26.7.so
-libresolv-2.26.3.so
-librt-2.26.3.4.so
-libselinux.so.1.1.4.1
-libsystemd.so.0.19.0.4.2
-libtinfo.so.5.9.4.2
-libudev.so.1.6.6.4.7
-libwrap.so.0.7.6.1.5
-libz.so.1.2.1.1.2.4
