+++ forever/config.js
@@ -1,5 +1,4 @@
 const config = {
     nodiscard: true
 }
-
 module.exports = config;
