+++ network/delete
@@ -1,0 +1,22 @@
+#!/usr/bin/env bash
+
+# This script deletes the network namespace for the given bot id
+
+botid=$1
+
+NS="catbotns$botid"
+VETH="veth$botid"
+VPEER_ADDR="10.200.$botid.2"
+INTERFACE=$(route -n | grep '^0\.0\.0\.0' | grep -o '[^ ]*$' | head -1)
+
+if [[ $EUID -ne 0 ]]; then
+  echo "You must be root to run this script!"
+  exit 1
+fi
+
+ip netns delete "$NS"
+ip link del "$VETH"
+
+iptables -t nat -D POSTROUTING -s "${VPEER_ADDR}"/24 -o "$INTERFACE" -j MASQUERADE
+iptables -D FORWARD -i "$INTERFACE" -o "${VETH}" -j ACCEPT
+iptables -D FORWARD -o "$INTERFACE" -i "${VETH}" -j ACCEPT
