+++ network/route
@@ -1,0 +1,49 @@
+#!/usr/bin/env bash
+
+# This script creates a network namespace for the given bot id
+
+botid=$1
+
+NS="catbotns$botid"
+VETH="veth$botid"
+VPEER="vpeer$botid"
+VETH_ADDR="10.200.$botid.1"
+VPEER_ADDR="10.200.$botid.2"
+INTERFACE=$(route -n | grep '^0\.0\.0\.0' | grep -o '[^ ]*$' | head -1)
+
+if [[ $EUID -ne 0 ]]; then
+  echo "You must be root to run this script!"
+  exit 1
+fi
+
+# Remove namespace if it exists.
+ip netns del "$NS" &>/dev/null
+ip link del "${VETH}" &>/dev/null
+
+# Create namespace
+ip netns add "$NS"
+
+# Create veth link.
+ip link add "${VETH}" type veth peer name "${VPEER}"
+
+# Add peer-1 to NS.
+ip link set "${VPEER}" netns "$NS"
+
+# Setup IP address of ${VETH}.
+ip addr add "${VETH_ADDR}"/24 dev "${VETH}"
+ip link set "${VETH}" up
+
+# Setup IP ${VPEER}.
+ip netns exec "$NS" ip addr add "${VPEER_ADDR}"/24 dev "${VPEER}"
+ip netns exec "$NS" ip link set "${VPEER}" up
+ip netns exec "$NS" ip link set lo up
+ip netns exec "$NS" ip route add default via "${VETH_ADDR}"
+
+# Enable IP-forwarding.
+echo 1 >/proc/sys/net/ipv4/ip_forward
+
+# Enable masquerading of 10.200.1.0.
+iptables -t nat -A POSTROUTING -s "${VPEER_ADDR}"/24 -o "$INTERFACE" -j MASQUERADE
+
+iptables -A FORWARD -i "$INTERFACE" -o "${VETH}" -j ACCEPT
+iptables -A FORWARD -o "$INTERFACE" -i "${VETH}" -j ACCEPT
