+++ package.json
@@ -8,30 +8,30 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/nullifiedcat/cathook-ipc-web-panel.git"
+    "url": "git+https://github.com/explowz/cathook-ipc-web-panel.git"
   },
-  "author": "nullifiedcat",
+  "author": "nullifiedcat & Bintr",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/nullifiedcat/cathook-ipc-web-panel/issues"
+    "url": "https://github.com/explowz/cathook-ipc-web-panel/issues"
   },
-  "homepage": "https://github.com/nullifiedcat/cathook-ipc-web-panel#readme",
+  "homepage": "https://github.com/explowz/cathook-ipc-web-panel#readme",
   "dependencies": {
-    "body-parser": "^1.19.0",
+    "body-parser": "^1.20.1",
     "browser-request": "^0.3.3",
     "browserify": "^17.0.0",
-    "express": "^4.17.3",
-    "express-session": "^1.16.1",
+    "express": "^4.18.2",
+    "express-session": "^1.17.3",
     "extend": "^3.0.2",
-    "format-duration": "^1.3.1",
-    "jquery": "^3.5.1",
+    "format-duration": "^2.0.0",
+    "jquery": "^3.6.1",
     "npid": "^0.4.0",
     "procfs-stats": "^1.0.2",
-    "randomstring": "^1.1.5",
-    "request": "^2.88.0",
+    "randomstring": "^1.2.3",
+    "request": "^2.88.2",
     "stoppable": "^1.1.0",
-    "tail": "^2.1.1",
+    "tail": "^2.2.4",
     "time-stamp": "^2.2.0",
-    "underscore": "^1.12.1"
+    "underscore": "^1.13.6"
   }
 }
