--- watchify.sh
@@ -1,1 +1,0 @@
-watchify script.js -o public/bundle.js
