+++ wpanel/design/style.css
@@ -1,0 +1,74 @@
+/* table fix */
+tr {
+    border-top: 1px solid transparent !important;
+    padding: 40px !important;
+}
+
+/* button fix */
+.lesshigh {
+    font-size: 9px;
+    height: 1.3rem;
+}
+
+button {
+    color: white;
+}
+
+/* rainbows */
+.rainbow {
+    background-image: repeating-linear-gradient(45deg, violet, indigo, blue, green, yellow, orange, red, violet);
+    text-align: center;
+    background-size: 800% 800%;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    font-size: 23px;
+    animation: rainbow 8s ease infinite;
+}
+
+@keyframes rainbow {
+    0% {
+        background-position: 0 50%
+    }
+    50% {
+        background-position: 100% 25%
+    }
+    100% {
+        background-position: 0 50%
+    }
+}
+
+/* selection */
+::selection {
+    background: #007bff;
+    color: white;
+}
+
+/* fix */
+#spanfix {
+    width: 6rem !important;
+}
+
+/* cursor's */
+button, a:hover, .btn {
+    cursor: url('cursor/hover.cur'), auto !important;
+}
+
+/* width */
+::-webkit-scrollbar {
+    width: 2px;
+}
+
+/* Track */
+::-webkit-scrollbar-track {
+    background: transparent;
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+    background: #888;
+}
+
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+    background: #555;
+}
